+++ example/lib/profile_screen/controller/profile_controller.dart
@@ -1,5 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ProfileController extends ChangeNotifier {
-  // ProfileController logic here
-}