+++ example/lib/profile_screen/repository/profile_repo.dart
@@ -1,5 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ProfileRepo {
-  // ProfileRepo logic here
-}