+++ example/lib/profile_screen/repository/profile_repo_interface.dart
@@ -1,5 +1,0 @@
-import 'package:flutter/material.dart';
-
-abstract class ProfileRepoInterface {
-  // ProfileRepoInterface logic here
-}